Fix sonar critical and major code smells

+++ src/items/decay/decay.cpp
@@ -43,11 +43,11 @@ void Decay::startDecay(Item* item)
 
 		int64_t timestamp = OTSYS_TIME() + duration;
 		if (decayMap.empty()) {
-			eventId = g_scheduler().addEvent(createSchedulerTask(std::max<int32_t>(SCHEDULER_MINTICKS, duration), std::bind_front(&Decay::checkDecay, this)));
+			eventId = g_scheduler().addEvent(createSchedulerTask(std::max<int32_t>(SCHEDULER_MINTICKS, static_cast<const int32_t>(duration)), std::bind_front(&Decay::checkDecay, this)));
 		} else {
 			if (timestamp < decayMap.begin()->first) {
 				g_scheduler().stopEvent(eventId);
-				eventId = g_scheduler().addEvent(createSchedulerTask(std::max<int32_t>(SCHEDULER_MINTICKS, duration), std::bind_front(&Decay::checkDecay, this)));
+				eventId = g_scheduler().addEvent(createSchedulerTask(std::max<int32_t>(SCHEDULER_MINTICKS, static_cast<const int32_t>(duration)), std::bind_front(&Decay::checkDecay, this)));
 			}
 		}
